+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 # Dev
 - added in skip rounds for masking and selfcal
-
+- Basic handling of CASDA measurement sets (preprocessing)
+- 
 ## 0.2.4
 
 - Added new masking modes